django/requirements.txt
asgiref==3.8.1
Django==5.1
sqlparse==0.5.1
django-environ==0.11.2
requests==2.32.3
PyJWT==2.9.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers == 4.4.0
pyotp==2.9.0
pip-system-certs==4.0
qrcode==7.4.2
django-two-factor-auth==1.17.0
pytz==2023.3
psycopg==3.1.18
